requirements.txt
pandas>=2.0.0
litellm>=1.51.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
tabulate>=0.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
ollama>=0.3.0
groq>=0.9.0
openai>=1.40.0
pypdf>=5.0.0
